pyproject.toml
[project]
name = "abac-sak"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask==2.3.2",
    "google-genai>=1.55.0",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scipy>=1.16.3",
]
